Moved FadePageTransformer and PagerAdapter into separate files because they where exactly the same in AppIntro and AppIntro2
