+++ library/src/main/java/com/github/paolorotolo/appintro/PagerAdapter.java
@@ -1,0 +1,30 @@
+package com.github.paolorotolo.appintro;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+import java.util.List;
+
+class PagerAdapter extends FragmentPagerAdapter {
+    private List<Fragment> fragments;
+
+    public PagerAdapter(FragmentManager fm, List<Fragment> fragments) {
+        super(fm);
+        this.fragments = fragments;
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        return this.fragments.get(position);
+    }
+
+    @Override
+    public int getCount() {
+        return this.fragments.size();
+    }
+
+    public List<Fragment> getFragments() {
+        return fragments;
+    }
+}
